Remove repeated options.js functions
