+++ src/main/node_options.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const path = require("path");
+const inferParser = require("./infer-parser");
 const getSupportInfo = require("../common/support").getSupportInfo;
 const normalizer = require("./options-normalizer");
 const loadPlugins = require("../common/load-plugins");
@@ -97,24 +97,6 @@ function normalize(options, opts) {
     supportOptions,
     Object.assign({ passThrough: Object.keys(hiddenDefaults) }, opts)
   );
-}
-
-function inferParser(filepath, plugins) {
-  const extension = path.extname(filepath);
-  const filename = path.basename(filepath).toLowerCase();
-
-  const language = getSupportInfo(null, {
-    plugins,
-    pluginsLoaded: true
-  }).languages.find(
-    language =>
-      language.since !== null &&
-      (language.extensions.indexOf(extension) > -1 ||
-        (language.filenames &&
-          language.filenames.find(name => name.toLowerCase() === filename)))
-  );
-
-  return language && language.parsers[0];
 }
 
 module.exports = { normalize, hiddenDefaults };
