+++ src/main/options.js
@@ -1,9 +1,8 @@
 "use strict";
 
-const path = require("path");
+const inferParser = require("./infer-parser");
 const getSupportInfo = require("../common/support").getSupportInfo;
 const normalizer = require("./options-normalizer");
-const assert = require("assert");
 const resolveParser = require("./parser").resolveParser;
 const getPlugin = require("./get-plugin");
 
@@ -95,24 +94,6 @@ function normalize(options, opts) {
     supportOptions,
     Object.assign({ passThrough: Object.keys(hiddenDefaults) }, opts)
   );
-}
-
-function inferParser(filepath, plugins) {
-  const extension = path.extname(filepath);
-  const filename = path.basename(filepath).toLowerCase();
-
-  const language = getSupportInfo(null, {
-    plugins,
-    pluginsLoaded: true
-  }).languages.find(
-    language =>
-      language.since !== null &&
-      (language.extensions.indexOf(extension) > -1 ||
-        (language.filenames &&
-          language.filenames.find(name => name.toLowerCase() === filename)))
-  );
-
-  return language && language.parsers[0];
 }
 
 module.exports = { normalize, hiddenDefaults };
